requirements.txt
transformers>=4.30.0
sentencepiece>=0.1.99
protobuf>=3.20.0
torch>=2.0.0
typing_extensions>=4.5.0
biopython>=1.80
h5py>=3.0.0
tqdm>=4.65.0
